requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
easyocr==1.7.0
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-multipart==0.0.6
Pillow==10.0.1
pydantic==2.5.0
python-levenshtein==0.21.1
jsonschema==4.19.2
aiofiles==23.2.1
PyPDF2==3.0.1
pdf2image==1.16.3
